[asan/mac] Fix a -Wformat warning
internal_getpid() is manually cast to int in all other places that
pass its result to a printf string, so do that here too for now.

+++ compiler-rt/lib/sanitizer_common/sanitizer_symbolizer_mac.cpp
@@ -68,7 +68,7 @@ class AtosSymbolizerProcess final : public SymbolizerProcess {
   bool StartSymbolizerSubprocess() override {
     // Put the string command line argument in the object so that it outlives
     // the call to GetArgV.
-    internal_snprintf(pid_str_, sizeof(pid_str_), "%d", internal_getpid());
+    internal_snprintf(pid_str_, sizeof(pid_str_), "%d", (int)internal_getpid());
 
     // Configure sandbox before starting atos process.
     return SymbolizerProcess::StartSymbolizerSubprocess();
